[weather.noaa] 3.0.1
